refactor(admin): extract Starter Prompts section from AppFormEditor (#1781)
Continues the incremental split of AppFormEditor.jsx proposed in #1781,
following the same pattern as the Upload Configuration (#1942) and
Variables (#2044) extractions. Moves the starter-prompt list editor and
its three self-contained immutable-update handlers (handleStarterPromptChange,
addStarterPrompt, removeStarterPrompt) into a new
client/src/features/admin/components/app-form/StarterPromptsSection.jsx,
since they only ever touch app.starterPrompts. AppFormEditor.jsx shrinks
from 2902 to 2786 lines. No behavior change intended — JSX and handler
logic ported verbatim.
